requirements.txt
streamlit==1.24.0
langchain==0.0.221
python-dotenv
openai
